Allow decimal number grades and extra credit and disallow negative grades

+++ backend/pkg/models/grade.go
@@ -3,18 +3,18 @@ package models
 import "fmt"
 
 type Grade struct {
-	StudentID   string `firestore:"studentID"`
-	Grade       int    `firestore:"grade"`
-	GradedBy    string `firestore:"gradedBy"`
-	TimeUpdated string `firestore:"timeUpdated"`
+	StudentID   string  `firestore:"studentID"`
+	Grade       float32 `firestore:"grade"`
+	GradedBy    string  `firestore:"gradedBy"`
+	TimeUpdated string  `firestore:"timeUpdated"`
 }
 
 type CreateGradeRequest struct {
-	CourseID     string `json:"courseid,omitempty"`
-	AssignmentID string `json:"assignmentid,omitempty"`
-	StudentID    string `json:"studentid"`
-	Grade        int    `json:"grade"`
-	GradedBy     *User  `json:"gradedBy,omitempty"`
+	CourseID     string  `json:"courseid,omitempty"`
+	AssignmentID string  `json:"assignmentid,omitempty"`
+	StudentID    string  `json:"studentid"`
+	Grade        float32 `json:"grade"`
+	GradedBy     *User   `json:"gradedBy,omitempty"`
 }
 
 type DeleteGradeRequest struct {
